+++ src/Knp/Menu/ItemInterface.php
@@ -448,20 +448,19 @@ interface ItemInterface extends  \ArrayAccess, \Countable, \IteratorAggregate
     function getPathAsString($separator = ' > ');
 
     /**
-     * Renders an array of label => uri pairs ready to be used for breadcrumbs.
+     * Renders an array ready to be used for breadcrumbs.
+     *
+     * Each element in the array will be an array with 3 keys:
+     * - `label` containing the label of the item
+     * - `url` containing the url of the item (may be `null`)
+     * - `item` containing the original item (may be `null` for the extra items)
      *
      * The subItem can be one of the following forms
      *   * 'subItem'
+     *   * Knp\Menu\ItemInterface object
      *   * array('subItem' => '@homepage')
      *   * array('subItem1', 'subItem2')
-     *
-     * @example
-     * // drill down to the Documentation menu item, then add "Chapter 1" to the breadcrumb
-     * $arr = $menu['Documentation']->getBreadcrumbsArray('Chapter 1');
-     * foreach ($arr as $name => $url)
-     * {
-     *
-     * }
+     *   * array(array('label' => 'subItem1', 'url' => '@homepage'), array('label' => 'subItem2'))
      *
      * @param  mixed $subItem A string or array to append onto the end of the array
      * @return array
@@ -472,6 +471,8 @@ interface ItemInterface extends  \ArrayAccess, \Countable, \IteratorAggregate
      * Returns the current menu item if it is a child of this menu item
      *
      * @return \Knp\Menu\ItemInterface|null
+     * @deprecated this method is flawed and will be removed in 2.0
+     * @see \Knp\Menu\Iterator\CurrentItemFilterIterator
      */
     function getCurrentItem();