+++ src/Knp/Menu/Iterator/CurrentItemFilterIterator.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Knp\Menu\Iterator;
+
+/**
+ * Filter iterator keeping only current items
+ */
+class CurrentItemFilterIterator extends \FilterIterator
+{
+    public function accept()
+    {
+        return $this->current()->isCurrent();
+    }
+}