+++ src/Knp/Menu/Iterator/DisplayedItemFilterIterator.php
@@ -1,0 +1,19 @@
+<?php
+
+namespace Knp\Menu\Iterator;
+
+/**
+ * Filter iterator keeping only current items
+ */
+class DisplayedItemFilterIterator extends \RecursiveFilterIterator
+{
+    public function accept()
+    {
+        return $this->current()->isDisplayed();
+    }
+
+    public function hasChildren()
+    {
+        return $this->current()->getDisplayChildren() && parent::hasChildren();
+    }
+}